reorganized filtering directory structure
